The desktop day fills the glass, and stacked saints open into a constellation
The chrome runs the width of the window and its four items are twice the
size. The theme toggle is the one thing not doubled: it is a box around a
glyph, and 64px of it beside 27px type reads as a button that has escaped
rather than a bigger one. The bar is 58.14px tall for it, which
--chrome-h-reserve now says at a third breakpoint — and note that the test
pinning it cannot catch this change being backed out, the reservation being
a min-height that holds the bar open however small the type inside it. The
one that does catch it is new and measures the ratio to the narrow sizes.

The day's own column takes the slack now: main gives up 108ch for the
window, and the *right* column is what is held rather than both sharing the
surplus, because the readings are four references and do not want a wider
measure. The week and month moved to the top of that column, which is what
sets its floor at 25rem: seven days, two shoulders and the month toggle need
about 400px, and at 330 the rail showed six days and the month would not
unfurl at all. The picker takes the phone's own 24px peek there for the
phone's own reason — a narrow rail is scrolled rather than read across — and
it has to take the phone's *gaps* with it: copying the three custom
properties and leaving those behind put the month's day names 3px off the
week's, which is a test this repository already had.

Two things the reader will notice and no instruction asked for, both
consequences: the month unfurls beside the day rather than pushing it down,
and the gold rule under the date spans the left column rather than the page.

The card is at least eighteen lines of its own preview tall and the
picture's column is derived from that height, which is what stops a wider
column making portrait icons enormous while it makes landscape ones right:
Lupus at 1:1.6 in a 414px column is 662px of icon over a 505px card. Because
it is a minimum the browser test pins the resolved min-height rather than the
rendered one — the card is usually taller, so an at-least assertion would
pass with the rule deleted. The month prints in full again, reversing
2026-08-24's abbreviation, which was bought when the heading shared a narrow
column.

And "...continue reading >" ends the preview, at both widths. On a phone
there is no preview for it to end — the lede needs 760px — so it sits under
the dates instead. Unlike the picture beside it, it is not hidden from a
screen reader: the picture is a wordless second link to a page the name
already opens, this has words and a different promise, and it says whose
life it opens.

On the map, saints recorded at one coordinate are spread into a tight ring
measured in degrees. That is the third arrangement in a day and the unit is
why it is not the first: the fan was screen pixels, so it covered more
country the further out the reader went and never resolved; this is
sub-pixel with the world on screen, so the merge still collapses it into one
honest mark, and it opens into a constellation about ten pixels across only
at the zooms the ceiling was raised to reach. Rounded under Mercator by
squashing the latitude, or it draws as an ellipse at Kyiv and worse further
north.

+++ src/views/daily/format.js
@@ -11,11 +11,15 @@ import { formatDate } from '../../lib/i18n.js';
  * because it reads as prose where the heading reads as a label.
  */
 
-// The page's own date wears the abbreviated month (author, 2026-08-24:
-// "display abbreviated months, e.g. Aug"); the buttons' aria-labels keep
-// dayFmt's full month, because a label is spoken, not glanced at.
+// The page's own date wears the month in full again (author, 2026-09-01: "on
+// Daily page print the full month name"), reversing 2026-08-24's "display
+// abbreviated months, e.g. Aug". The abbreviation was bought when the heading
+// shared its line with a narrow column; the desktop page is two columns and
+// most of a window wide now, and "1 Sept 2026" was saving room nothing needs.
+// The month gutter beside the grid keeps its own abbreviation — `monthFmt`
+// below — where the reason still holds.
 export const headingFmt = (d) =>
-  formatDate({ weekday: 'long', day: 'numeric', month: 'short', year: 'numeric', timeZone: 'UTC' }, d);
+  formatDate({ weekday: 'long', day: 'numeric', month: 'long', year: 'numeric', timeZone: 'UTC' }, d);
 
 export const weekdayFmt = (d) => formatDate({ weekday: 'short', timeZone: 'UTC' }, d);
